// src/app/projects/kmmp/actividades/models/bays-model.ts
export interface BayI {
  nombre: string;
  actividad: string;
  tipoMantenimiento?: string;
  documento?: string;
  ruc?: string;
  razon?: string;
  ubicacion?: string;
  nestado: string;
  tag?: string;
  modelo: string;
  flota: string;
  cliente: string;
  horometro?: string;
  idCliente: number;
  idActividad: number;
  idTipoMantenimiento: number;
  idModelo: number;
  key?: number;
  rownum?: number;
  total?: number;
  id: number;
  codigo?: string;
  descripcion?: string;
  usuario?: string;
  estado?: number;
  fechaReg?: string;
  fechaMod?: string;
}